requirements.txt
asgiref==3.3.1
boto3==1.17.67
botocore==1.20.67
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
crispy-forms-gds==0.2.2
cryptography==3.4.7
cssselect2==0.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.3
django-admin-thumbnails==0.2.5
django-allauth==0.44.0
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.2.1
django-import-export==2.5.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-pandas==0.6.6
django-simple-coupons==0.5
django-weasyprint==1.1.0
et-xmlfile==1.0.1
gunicorn==20.1.0
html5lib==1.1
idna==2.10
jmespath==0.10.0
MarkupPy==1.14
numpy==1.23.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.9
pandas==1.4.3
Pillow==9.3.0
product==0.1.1.1
pycairo==1.20.1
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
pyphen==0.11.0
python-dateutil==2.8.1
python-gettext==4.0
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.4.2
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
tinycss2==1.1.0
tk==0.1.0
urllib3==1.26.4
vendor==0.1
views==0.3
WeasyPrint==52.5
webencodings==0.5.1
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
